refactor: inline screens directly into expo-router routes
Move each screen's code straight into its app/ route file instead of
re-exporting from src/screens, per Expo Router convention. Relocates
supporting helpers (form hooks, authStyles, profile mock data, and the
Profile Recipe/Saved tabs) out of src/screens into src/hooks,
src/constants, and src/components/Profile respectively, and removes
src/screens entirely.
