+++ src/constants/profileData.tsx
@@ -4,119 +4,119 @@ export const data = [
         profile_name: 'Tadeu Melembe',
         created_at: '2h ago',
         title: 'Red Wine and Mint Soufflé',
-        image: require('../../../assets/images/dish-2.png'),
+        image: require('../../assets/images/dish-2.png'),
         description: 'Integer at bibendum nisi. Integer eu tortor orci. Proin in consequat odio, vitae viverra est. Integer neque mauris, consequat vitae sapien sit amet, pretium convallis velit. Donec auctor velit consectetur rutrum hendrerit. Nam ornare vel urna quis dictum. Sed feugiat, magna eget ullamcorper scelerisque, nulla urna consequat magna, scelerisque tempus sem ex in ex. Sed a sagittis ante. Aenean non eros a urna sodales aliquam. Nulla risus lectus, accumsan dignissim nunc ac, placerat ullamcorper nibh.'
     },
     {
         id: '2',
         profile_name: 'Wildson Guiamba',
         created_at: '1day ago',
-        image: require('../../../assets/images/dish-1.png'),
+        image: require('../../assets/images/dish-1.png'),
         title: 'White Wine Toffee',
         description: 'Proin sit amet dictum turpis, at aliquam nulla. Cras tincidunt, nisl quis tincidunt hendrerit, orci lorem pharetra nunc, sed aliquam eros nunc quis purus. Suspendisse vulputate nisl vitae est finibus, id tristique dui suscipit. In condimentum eros nisl, ut interdum turpis suscipit vitae. Suspendisse hendrerit pulvinar lacus, in egestas risus gravida ut.'
     },
     {
         id: '3',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '334',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '343',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '3434',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '322',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '312323',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '3435',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '8863',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '1334',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '23346',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '1121',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '112212',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '31233',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '35421',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     },
     {
         id: '863',
         profile_name: 'Itan',
-        image: require('../../../assets/images/dish-3.png'),
+        image: require('../../assets/images/dish-3.png'),
         created_at: '50min ago',
         title: 'Vanilla Pud',
     }
