--- src/screens/Auth/Login/index.tsx
@@ -1,1 +1,0 @@
-export {default} from './LoginScreen'
