--- src/screens/Auth/SignUp/index.tsx
@@ -1,1 +1,0 @@
-export {default} from './SignUpScreen'
