--- src/screens/Profile/index.tsx
@@ -1,1 +1,0 @@
-export {default} from './ProfileScreen'
