--- src/screens/Settings/index.tsx
@@ -1,1 +1,0 @@
-export {default} from './SettingsScreen'
