--- src/screens/TabHome/index.tsx
@@ -1,1 +1,0 @@
-export {default} from './HomeScreen'
