--- src/screens/TabSearch/index.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import { ScrollView, Text } from "../../components/Themed";
-import { Screen } from "../../presentation/components/ui/Screen";
-import style from "../../constants/style";
-
-const TabHome: React.FC = () => {
-    return (
-        <Screen style={style.horizontalPadding}>
-            <ScrollView>
-                <Text>Search</Text>
-            </ScrollView>
-        </Screen>
-    )
-}
-
-export default TabHome
